Build files for Mac
git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@11887 222415a0-0dc2-4184-858d-a27cde375526

Former-commit-id: 64adf60033bf135fca7a93c5385b9717f58ed7b6

+++ plugins/DataInterfacePlugin/src/base/datareader/FileReader.cpp
@@ -196,7 +196,7 @@ bool FileReader::ParseRealValue(const Integer i, const std::string& theField)
       do
       {
          ++end;
-      } while ((theLine[end] != ' ') && (end < (Integer)theLine.length()));
+      } while ((theLine[end] != ' ') && (end < theLine.length()));
 
       std::string theData = theLine.substr(start, end - start);
